Move header_proxy_protocol and proxy_protocol to network/protocol/proxy_protocol
