+++ lib/src/network/mod.rs
@@ -12,7 +12,6 @@ pub mod protocol;
 pub mod http;
 pub mod backends;
 pub mod retry;
-pub mod header_proxy_protocol;
 
 #[cfg(feature = "splice")]
 mod splice;
